+++ integration-tests/pytest.ini
@@ -6,6 +6,7 @@ testpaths =
     messaging
     connectors
 # storage skip for now
+    agents
     enterprise_search
     response-validation
